+++ src/nodes/NodeManager.ts
@@ -14,7 +14,7 @@ import type {
 } from '../nodes/types';
 import type { ClaimEncoded } from '../claims/types';
 import type { Timer } from '../types';
-import type { PromiseDeconstructed } from '../utils/utils';
+import type { PromiseDeconstructed } from '../types';
 import Logger from '@matrixai/logger';
 import { StartStop, ready } from '@matrixai/async-init/dist/StartStop';
 import * as nodesErrors from './errors';
